split model interfaces from communityApiService.ts to their own modules
