+++ server/models/delius/deliusUser.ts
@@ -1,0 +1,13 @@
+export interface DeliusRole {
+  name: string
+}
+
+export default interface DeliusUser {
+  userId: string
+  username: string
+  firstName: string
+  surname: string
+  email: string
+  enabled: boolean
+  roles: Array<DeliusRole>
+}
