+++ server/routes/referrals/referralsController.ts
@@ -29,7 +29,8 @@ import CheckAnswersView from './checkAnswersView'
 import CheckAnswersPresenter from './checkAnswersPresenter'
 import ConfirmationView from './confirmationView'
 import ConfirmationPresenter from './confirmationPresenter'
-import CommunityApiService, { DeliusServiceUser } from '../../services/communityApiService'
+import CommunityApiService from '../../services/communityApiService'
+import DeliusServiceUser from '../../models/delius/deliusServiceUser'
 import errorMessages from '../../utils/errorMessages'
 import logger from '../../../log'
 import ServiceUserDetailsPresenter from './serviceUserDetailsPresenter'
