+++ server/routes/shared/submittedPostSessionFeedbackPresenter.ts
@@ -1,4 +1,4 @@
-import { DeliusServiceUser } from '../../services/communityApiService'
+import DeliusServiceUser from '../../models/delius/deliusServiceUser'
 import AuthUser from '../../models/authUser'
 import { ActionPlanAppointment } from '../../models/actionPlan'
 import DateUtils from '../../utils/dateUtils'
