+++ server/services/interventionsService.test.ts
@@ -10,7 +10,7 @@ import serviceCategoryFactory from '../../testutils/factories/serviceCategory'
 import serviceProviderFactory from '../../testutils/factories/serviceProvider'
 import eligibilityFactory from '../../testutils/factories/eligibility'
 import interventionFactory from '../../testutils/factories/intervention'
-import { DeliusServiceUser } from './communityApiService'
+import DeliusServiceUser from '../models/delius/deliusServiceUser'
 import actionPlanFactory from '../../testutils/factories/actionPlan'
 import actionPlanAppointmentFactory from '../../testutils/factories/actionPlanAppointment'
 import endOfServiceReportFactory from '../../testutils/factories/endOfServiceReport'
