+++ server/services/interventionsService.ts
@@ -2,7 +2,7 @@ import RestClient from '../data/restClient'
 import logger from '../../log'
 import { ApiConfig } from '../config'
 import { SanitisedError } from '../sanitisedError'
-import { DeliusServiceUser } from './communityApiService'
+import DeliusServiceUser from '../models/delius/deliusServiceUser'
 import CalendarDay from '../utils/calendarDay'
 import AuthUser from '../models/authUser'
 import EndOfServiceReport, { AchievementLevel } from '../models/endOfServiceReport'
