+++ server/utils/controllerUtils.ts
@@ -1,7 +1,7 @@
 import { Response } from 'express'
 import LayoutPresenter from '../routes/shared/layoutPresenter'
 import LayoutView, { PageContentView } from '../routes/shared/layoutView'
-import { DeliusServiceUser } from '../services/communityApiService'
+import DeliusServiceUser from '../models/delius/deliusServiceUser'
 
 export default class ControllerUtils {
   static renderWithLayout(res: Response, contentView: PageContentView, serviceUser: DeliusServiceUser | null): void {
